requirements.txt
zeroentropy
python-dotenv
ollama
fastapi
uvicorn
